requirements.txt
requests~=2.32.3
Typer~=0.12.3
click==8.1.7
python-dotenv~=1.0.1